[refactor]: Domain 서비스 리팩토링 - 단일 책임 원칙 준수
**문제상황:**
Domain 계층의 서비스들이 generic한 *Service 네이밍을 사용하여 CLAUDE.md 가이드라인을 위반하고,
일부 서비스가 단일 책임 원칙을 위반하여 여러 책임을 가지고 있었음

**Solution:**
1. 모든 domain 서비스를 구체적이고 명시적인 이름으로 변경
2. UserPasswordService를 단일 책임에 따라 PasswordValidator와 PasswordEncoder로 분리
3. 각 서비스가 하나의 명확한 책임만 가지도록 리팩토링

**변경사항:**
- UserAuthenticationService → UserAuthenticator로 이름 변경
- SignUpPendingUserService → PendingUserValidator로 이름 변경
- UserPasswordService를 PasswordValidator와 PasswordEncoder로 분리
- 인프라스트럭처 계층에 SpringPasswordEncoder, SpringPasswordValidator 구현체 추가
- 모든 의존성 참조 및 테스트 코드 업데이트
- Fake 객체들도 동일한 네이밍 규칙으로 업데이트
