+++ service/user/src/main/kotlin/me/helloc/techwikiplus/user/domain/service/EmailTemplateGenerator.kt
@@ -1,12 +1,12 @@
 package me.helloc.techwikiplus.user.domain.service
 
 interface EmailTemplateGenerator {
-    fun generateVerificationEmail(code: String): EmailTemplate
+    fun generateVerificationEmail(code: String): EmailTemplateDetails
 
-    fun generatePasswordResetEmail(code: String): EmailTemplate
+    fun generatePasswordResetEmail(code: String): EmailTemplateDetails
+
+    data class EmailTemplateDetails(
+        val subject: String,
+        val body: String,
+    )
 }
-
-data class EmailTemplate(
-    val subject: String,
-    val body: String,
-)
