+++ service/user/src/main/kotlin/me/helloc/techwikiplus/user/infrastructure/passwordencoder/SpringPasswordEncoder.kt
@@ -1,13 +1,13 @@
 package me.helloc.techwikiplus.user.infrastructure.passwordencoder
 
-import me.helloc.techwikiplus.user.domain.service.UserPasswordService
-import org.springframework.security.crypto.password.PasswordEncoder
+import me.helloc.techwikiplus.user.domain.service.PasswordEncoder
 import org.springframework.stereotype.Component
+import org.springframework.security.crypto.password.PasswordEncoder as SpringPasswordEncoderInterface
 
 @Component
-class SpringPasswordService(
-    val passwordEncoder: PasswordEncoder,
-) : UserPasswordService {
+class SpringPasswordEncoder(
+    private val passwordEncoder: SpringPasswordEncoderInterface,
+) : PasswordEncoder {
     override fun encode(password: String): String {
         return passwordEncoder.encode(password)
     }
