+++ service/user/src/main/kotlin/me/helloc/techwikiplus/user/infrastructure/passwordencoder/SpringPasswordValidator.kt
@@ -1,0 +1,7 @@
+package me.helloc.techwikiplus.user.infrastructure.passwordencoder
+
+import me.helloc.techwikiplus.user.domain.service.PasswordValidator
+import org.springframework.stereotype.Component
+
+@Component
+class SpringPasswordValidator : PasswordValidator
